Move patchH5vccCanvasKit from common.dart to h5vcc_test.dart since it's only used there.
Remove test in initialization_test.dart.
