+++ lib/web_ui/test/canvaskit/h5vcc_test.dart
@@ -2,6 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+import 'dart:js' as js;
+
 import 'package:test/bootstrap/browser.dart';
 import 'package:test/test.dart';
 import 'package:ui/src/engine.dart';
@@ -40,4 +42,38 @@ void testMain() {
       expect(getH5vccSkSurfaceCalledCount, 1);
     });
   }, testOn: 'chrome');
+}
+
+class PatchedH5vcc implements H5vcc {
+  @override
+  final CanvasKit canvasKit;
+
+  PatchedH5vcc(this.canvasKit);
+}
+
+/// Test setup to initialize `window.h5vcc` with a patched H5vcc implementation.
+///
+/// The patched H5vcc implementation uses a downloaded CanvasKit implementation
+/// monkey-patched with a fake getH5vccSkSurface function.
+void patchH5vccCanvasKit({Function? onGetH5vccSkSurfaceCalled}) {
+  CanvasKit? existingCanvasKit;
+  SkiaFontCollection? existingSkiaFontCollection;
+
+  setUpAll(() async {
+    // Set `window.h5vcc` to PatchedH5vcc which uses a downloaded CanvasKit.
+    final CanvasKit downloadedCanvasKit = await downloadCanvasKit();
+    debugH5vccSetter = PatchedH5vcc(downloadedCanvasKit);
+
+    // Monkey-patch the getH5vccSkSurface function of `window.h5vcc.canvasKit`.
+    js.context['h5vcc']['canvasKit']['getH5vccSkSurface'] = () {
+      if (onGetH5vccSkSurfaceCalled != null) {
+        onGetH5vccSkSurfaceCalled();
+      }
+
+      // Returns a fake [SkSurface] object with a minimal implementation.
+      return js.JsObject.jsify(<String, dynamic>{
+        'dispose': () {}
+      });
+    };
+  });
 }
