+++ lib/web_ui/test/canvaskit/initialization_test.dart
@@ -25,9 +25,5 @@ void testMain() {
       expect(html.document.body!.attributes['flt-build-mode'], 'debug');
     });
     // TODO(hterkelsen): https://github.com/flutter/flutter/issues/60040
-
-    test('does not include getH5vccSkSurface', () {
-      expect(canvasKit.getH5vccSkSurface, isNull);
-    }, testOn: 'chrome');
   }, skip: isIosSafari);
 }
